requirements.txt
python-rtmidi==1.1.2